requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.11.0
Authlib==1.4.1
python-dotenv==1.1.0
gunicorn==23.0.0
email-validator==2.2.0
requests==2.32.3
google-api-python-client==2.166.0
google-auth-oauthlib==1.2.1
